V20.3.7

+++ shell/browser/ui/views/menu_model_adapter.cc
@@ -14,7 +14,7 @@ MenuModelAdapter::~MenuModelAdapter() = default;
 bool MenuModelAdapter::GetAccelerator(int id,
                                       ui::Accelerator* accelerator) const {
   ui::MenuModel* model = menu_model_;
-  size_t index = 0;
+  int index = 0;
   if (ui::MenuModel::GetModelAndIndexForCommandId(id, &model, &index)) {
     return static_cast<ElectronMenuModel*>(model)->GetAcceleratorAtWithParams(
         index, true, accelerator);
